+++ components/archive/impl/src/main/java/com/flipperdevices/archive/impl/composable/category/ComposableCategoryCard.kt
@@ -10,31 +10,34 @@ import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import androidx.lifecycle.viewmodel.compose.viewModel
 import com.flipperdevices.archive.impl.viewmodel.CategoryViewModel
+import com.flipperdevices.archive.model.CategoryType
 import com.flipperdevices.core.ui.theme.LocalPallet
+import tangle.viewmodel.compose.tangleViewModel
 
 @Composable
 fun ComposableCategoryCard(
-    modifier: Modifier = Modifier
+    modifier: Modifier = Modifier,
+    onOpenCategory: (CategoryType) -> Unit,
 ) {
     Card(
         modifier = modifier.padding(top = 14.dp, start = 14.dp, end = 14.dp, bottom = 2.dp)
     ) {
-        ComposableCategoryList()
+        ComposableCategoryList(onOpenCategory)
     }
 }
 
 @Composable
 private fun ComposableCategoryList(
-    categoryViewModel: CategoryViewModel = viewModel()
+    onOpenCategory: (CategoryType) -> Unit,
+    categoryViewModel: CategoryViewModel = tangleViewModel(),
 ) {
     val categories by categoryViewModel.getCategoriesFlow().collectAsState()
     val deletedCategory by categoryViewModel.getDeletedFlow().collectAsState()
 
     Column {
         categories.forEach {
-            ComposableCategoryItem(categoryItem = it, categoryViewModel = categoryViewModel)
+            ComposableCategoryItem(categoryItem = it, onOpenCategory = onOpenCategory)
         }
 
         Divider(
@@ -43,9 +46,6 @@ private fun ComposableCategoryList(
             color = LocalPallet.current.divider12
         )
 
-        ComposableCategoryItem(
-            categoryItem = deletedCategory,
-            categoryViewModel = categoryViewModel
-        )
+        ComposableCategoryItem(categoryItem = deletedCategory, onOpenCategory = onOpenCategory)
     }
 }