+++ components/archive/impl/src/main/java/com/flipperdevices/archive/impl/composable/category/ComposableCategoryItem.kt
@@ -14,8 +14,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.dp
 import com.flipperdevices.archive.impl.model.CategoryItem
-import com.flipperdevices.archive.impl.viewmodel.CategoryViewModel
-import com.flipperdevices.core.ui.ktx.LocalRouter
+import com.flipperdevices.archive.model.CategoryType
 import com.flipperdevices.core.ui.ktx.clickableRipple
 import com.flipperdevices.core.ui.theme.LocalPallet
 import com.flipperdevices.core.ui.theme.LocalTypography
@@ -24,14 +23,12 @@ import com.flipperdevices.core.ui.res.R as DesignSystem
 @Composable
 fun ComposableCategoryItem(
     categoryItem: CategoryItem,
-    categoryViewModel: CategoryViewModel,
-    modifier: Modifier = Modifier
+    modifier: Modifier = Modifier,
+    onOpenCategory: (CategoryType) -> Unit,
 ) {
-    val router = LocalRouter.current
-
     Row(
         modifier = modifier
-            .clickableRipple { categoryViewModel.onCategoryClick(router, categoryItem) }
+            .clickableRipple { onOpenCategory(categoryItem.categoryType) }
             .padding(horizontal = 12.dp),
         verticalAlignment = Alignment.CenterVertically
     ) {