+++ components/bridge/dao/api/src/main/java/com/flipperdevices/bridge/dao/api/model/FlipperKeyContent.kt
@@ -4,6 +4,7 @@ import android.os.Parcelable
 import kotlinx.parcelize.IgnoredOnParcel
 import kotlinx.parcelize.Parcelize
 import kotlinx.serialization.Serializable
+import kotlinx.serialization.Transient
 import java.io.ByteArrayInputStream
 import java.io.File
 import java.io.FileInputStream
@@ -37,7 +38,13 @@ sealed class FlipperKeyContent : Parcelable {
     }
 
     @Parcelize
-    data class InternalFile(val file: File) : FlipperKeyContent() {
+    @Serializable
+    data class InternalFile(val path: String) : FlipperKeyContent() {
+
+        @IgnoredOnParcel
+        @Transient
+        private val file = File(path)
+
         override fun openStream(): InputStream {
             return if (file.exists()) {
                 FileInputStream(file)