+++ components/bridge/dao/api/src/main/java/com/flipperdevices/bridge/dao/api/model/navigation/FlipperKeyPathType.kt
@@ -1,4 +1,4 @@
-package com.flipperdevices.uploader.api
+package com.flipperdevices.bridge.dao.api.model.navigation
 
 import android.os.Bundle
 import androidx.navigation.NavType
@@ -7,7 +7,7 @@ import com.flipperdevices.core.ktx.android.parcelable
 import kotlinx.serialization.decodeFromString
 import kotlinx.serialization.json.Json
 
-class FlipperKeyPathType : NavType<FlipperKeyPath>(isNullableAllowed = true) {
+class FlipperKeyPathType : NavType<FlipperKeyPath>(isNullableAllowed = false) {
     override fun get(bundle: Bundle, key: String): FlipperKeyPath? {
         return bundle.parcelable(key)
     }