+++ components/keyedit/impl/src/main/java/com/flipperdevices/keyedit/impl/viewmodel/processors/LimboKeyProcessor.kt
@@ -15,7 +15,6 @@ import com.flipperdevices.keyedit.impl.R
 import com.flipperdevices.keyedit.impl.model.EditableKey
 import com.flipperdevices.keyedit.impl.model.KeyEditState
 import com.flipperdevices.singleactivity.api.SingleActivityApi
-import com.github.terrakok.cicerone.Router
 import javax.inject.Inject
 
 private const val NOTIFICATION_DURATION_MS = 3 * 1000L
@@ -53,7 +52,7 @@ class LimboKeyProcessor @Inject constructor(
     override suspend fun onSave(
         editableKey: EditableKey.Limb,
         editState: KeyEditState.Editing,
-        router: Router
+        onEndAction: () -> Unit
     ) {
         val newPathUnfree = if (editState.name != null) {
             editableKey.notSavedFlipperKey.mainFile.path.copyWithChangedName(editState.name)