+++ components/nfceditor/sample/src/main/java/com/flipperdevices/nfceditor/sample/NfcEditorApplication.kt
@@ -5,13 +5,14 @@ import com.flipperdevices.core.di.ApplicationParams
 import com.flipperdevices.core.di.ComponentHolder
 import com.flipperdevices.nfceditor.sample.di.DaggerAppComponent
 import com.flipperdevices.nfceditor.sample.di.NfcEditorComponent
+import tangle.inject.TangleGraph
 import timber.log.Timber
 
 class NfcEditorApplication : Application() {
 
     override fun onCreate() {
         super.onCreate()
-        ComponentHolder.components += DaggerAppComponent.factory()
+        val appComponent = DaggerAppComponent.factory()
             .create(
                 context = this,
                 application = this,
@@ -20,6 +21,9 @@ class NfcEditorApplication : Application() {
                     version = "Sample"
                 )
             )
+
+        ComponentHolder.components += appComponent
+        TangleGraph.add(appComponent)
 
         Timber.plant(Timber.DebugTree())
         val shake2report = ComponentHolder.component<NfcEditorComponent>().shake2report.get()