+++ components/share/receive/src/main/java/com/flipperdevices/share/receive/helpers/FlipperKeyParserHelper.kt
@@ -51,7 +51,7 @@ class FlipperKeyParserHelper @Inject constructor(
                     folder = fileKey.extension,
                     nameWithExtension = fileKey.name
                 ),
-                FlipperKeyContent.InternalFile(fileKey)
+                FlipperKeyContent.InternalFile(fileKey.absolutePath)
             ),
             synchronized = false,
             deleted = false